requirements.txt
# pip install virtualenv
# python -m venv venv
# venv/Scripts/activate         # On Windows
# source myvenv/bin/activate    # On macOS/Linux

python-dotenv
python-telegram-bot